requirements/dev.txt
-r dist.txt
mock>=1.0.1,<2
django-nose>=1.4.3,<2
nosexcover>=1.0.10,<2
